requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.3.0
altgraph==0.17.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.2.0
beautifulsoup4==4.11.2
blinker==1.8.2
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.3.1
chromedriver-binary-auto==0.1.2
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cryptography==37.0.2
cyberdrop-dl==4.0.72
cycler==0.11.0
ffmpeg-python==0.2.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.13.0
idna==3.3
importlib-resources==5.12.0
Jinja2==3.1.4
joblib==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
lxml==4.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
myjdapi==1.1.6
numpy==1.22.4
outcome==1.1.0
packaging==23.0
pandas==1.4.2
pefile==2021.9.3
piexif==1.1.3
Pillow==9.2.0
protobuf==4.25.3
pyarrow==16.1.0
pycparser==2.21
pycryptodome==3.17
pydeck==0.9.1
Pygments==2.14.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0
referencing==0.35.1
requests==2.27.1
rich==13.3.1
rpds-py==0.18.1
scikit-learn==1.2.1
scipy==1.10.1
selenium==4.1.5
six==1.16.0
smmap==5.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
streamlit==1.34.0
tenacity==8.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tornado==6.4
tqdm==4.64.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
watchdog==4.0.0
webdriver-manager==3.5.4
wsproto==1.1.0
yarl==1.8.2
zipp==3.15.0
